fix: code smells in pnl-multipliers-calculator
