+++ src/pages/trader-tools/pnl-for-multipliers/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Hero, StyledHeader } from '../common/_style'
-import PnLMultipliersCalculatorSection from './_pnl-for-multipliers'
+import PnLMultipliersCalculatorSection from './_pnl_multipliers_calculator'
 import { Container, SEO } from 'components/containers'
 import { localize, WithIntl } from 'components/localization'
 import Layout from 'components/layout/layout'
